requirements.txt
fastapi
uvicorn
python-dotenv
google-generativeai
groq
streamlit
requests
pydantic
